handled authentication

+++ v1/item/views.py
@@ -1,16 +1,21 @@
 # Django imports
 from django.shortcuts import render
 from django.http import JsonResponse
+from django.utils import timezone
+import datetime
 # DRF imports
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
+from rest_framework import generics
 # Serializer Class imports
 from .serializers import CategorySerializer, ItemSerializer
 # Model Class imports
 from .models import Category, Item
-from v1.shop.models import Cafe
+from v1.shop.models import Shop
+#importing pagination
+from .paginations import SmallOffsetPagination, SmallPagesPagination
 
 # APIs for category and Item
 @api_view(['GET'])
@@ -32,12 +37,11 @@ def apiOverview(request):
 
 # methods for category 
 # categoryList
-@api_view(['GET'])
-@permission_classes((IsAuthenticated,))
-def categoryList(request):
-  categories = Category.objects.all().order_by('-id')
-  serializer = CategorySerializer(categories, many=True)
-  return Response(serializer.data)
+class CategoryList(generics.ListAPIView):
+  queryset = Category.objects.all()
+  serializer_class = CategorySerializer
+  permission_classes = [IsAuthenticated]
+  pagination_class = SmallPagesPagination
 
 # categoryDetail
 @api_view(['GET'])
@@ -50,19 +54,11 @@ def categoryDetail(request, pk):
 # categoryCreate
 @api_view(['POST'])
 @permission_classes((IsAuthenticated,))
-def categoryCreate(request):
-  user = request.user
-  user_id = Category(user = user)
-  cafe = request.user.cafe
-  cafe_id = Category(cafe = cafe)
-  # cafe = Cafe.objects.filter(user_id=user.id).get()
-  # cafe_id = Category(cafe = cafe)
-  print(cafe)
-  # context = {'user_id': user_id, 'cafe_id': cafe_id}
-  serializer = CategorySerializer(data = {'user_id': user_id, 'cafe_id': cafe_id})
+def categoryCreate(request):  
+  serializer = CategorySerializer(data=request.data)
   data={}
   if serializer.is_valid():
-    serializer.save()
+    serializer.save(user=request.user, shop=request.user.shop)
     data["success"] = "Category Has Been Created!"
     return Response(serializer.data, status=status.HTTP_201_CREATED)
   return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -73,7 +69,7 @@ def categoryCreate(request):
 def categoryUpdate(request, pk):
   category = Category.objects.get(id=pk)
   user = request.user
-  if category.created_by != user:
+  if category.user != user:
     return Response({'Response':"You dont have Permission to edit this data."})
   serializer = CategorySerializer(instance=category, data=request.data)
   if serializer.is_valid():
@@ -86,13 +82,14 @@ def categoryUpdate(request, pk):
 def categoryDelete(request, pk):
   category = Category.objects.get(id=pk)
   user = request.user
-  if category.created_by != user:
+  if category.user != user:
     return Response({'Response':"You dont have Permission to delete this data."})
   category.delete()
-  return Response('Category succsesfully delete!')
+  return Response('Category Has Been Deleted!')
 
-# methods for Items
+# Item List
 @api_view(['GET'])
+@permission_classes((IsAuthenticated,))
 def itemList(request):
   try:
     items = Item.objects.all().order_by('-id')
@@ -102,7 +99,15 @@ def itemList(request):
     serializer = ItemSerializer(items, many=True)
   return Response(serializer.data)
 
+class ItemList(generics.ListAPIView):
+  queryset = Item.objects.all()
+  serializer_class = ItemSerializer
+  permission_classes = [IsAuthenticated]
+  pagination_class = SmallPagesPagination
+
+# Single Item Details
 @api_view(['GET'])
+@permission_classes((IsAuthenticated,))
 def itemDetail(request, pk):
   try:
     item = Item.objects.get(id=pk)
@@ -113,33 +118,46 @@ def itemDetail(request, pk):
     serializer = ItemSerializer(item, many=False)
   return Response(serializer.data)
 
+# Item Create
 @api_view(['POST'])
+@permission_classes((IsAuthenticated,))
 def itemCreate(request):
   if request.method == "POST":
+    shop=request.user.shop
+    print(shop)
     serializer = ItemSerializer(data=request.data)
     data={}
     if serializer.is_valid():
-      serializer.save()
+      serializer.save(user=request.user, shop=request.user.shop, date_added=datetime.datetime.now())
       data["success"]="Create Successful"
       return Response(serializer.data, status=status.HTTP_201_CREATED)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+# Item Update
 @api_view(['POST'])
+@permission_classes((IsAuthenticated,))
 def itemUpdate(request, pk):
   try:
     item = Item.objects.get(id=pk)
   except Item.DoesNotExist:
     return Response(status=status.HTTP_404_NOT_FOUND)
   if request.method == "POST":
+    user = request.user
+    if item.user != user:
+      return Response({'Response':"You dont have Permission to edit this data."})
     serializer = ItemSerializer(instance=item, data=request.data)
     data={}
     if serializer.is_valid():
       serializer.save()
-      data["success"]="update Successful"
+      data["success"]="Item Has Been Updated"
       return Response (data=data)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+# Item Delete
 @api_view(['DELETE'])
+@permission_classes((IsAuthenticated,))
 def itemDelete(request, pk):
   try:
     item = Item.objects.get(id=pk)
@@ -147,11 +165,8 @@ def itemDelete(request, pk):
     return Response(status=status.HTTP_404_NOT_FOUND)
 
   if request.method == "DELETE":
-    operation = item.delete()
-    data={}
-    if operation:
-      data["success"]="Delete Successful"
-    else:
-      data["failure"]="Delete Failed"
-    return Response(data=data)
-
+    user = request.user
+    if item.user != user:
+      return Response({'Response':"You dont have Permission to delete this data."})
+    item.delete()
+    return Response('Category Has Been Deleted!')
